requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6